Filter options

+++ src/main/java/com/smartstay/smartstay/services/InvoiceDiscountService.java
@@ -56,7 +56,7 @@ public class InvoiceDiscountService {
             double discountPercentage = 0.0;
             if (discount.discountPercentage() != null) {
                 discountPercentage = Double.parseDouble(String.valueOf(discount.discountPercentage()));
-                discountAmount = (double) Math.round(((double) discountPercentage / 100) * invDiscount.getInvoiceAmount());
+                discountAmount = (double) Math.round((discountPercentage / 100) * invDiscount.getInvoiceAmount());
             }
             else if (discount.discountAmount() != null) {
                 discountAmount = discount.discountAmount();
